fix(monobank): widen invoice amount fields from int to long

+++ backend/src/FuelFlow.API/Features/Orders/SharedServices/Monobank/Models/MonobankInvoiceStatus.cs
@@ -4,8 +4,8 @@ public sealed class MonobankInvoiceStatus
 {
     public string InvoiceId { get; set; } = null!;
     public string Status { get; set; } = null!;
-    public int? Amount { get; set; }
-    public int? FinalAmount { get; set; }
+    public long? Amount { get; set; }
+    public long? FinalAmount { get; set; }
     public DateTime? CreatedDate { get; set; }
     public DateTime? ModifiedDate { get; set; }
     public string? FailureReason { get; set; }
@@ -15,7 +15,7 @@ public sealed class MonobankInvoiceStatus
 public sealed class MonobankCancelListItem
 {
     public string Status { get; set; } = null!;
-    public int? Amount { get; set; }
+    public long? Amount { get; set; }
     public int? Ccy { get; set; }
     public DateTime? CreatedDate { get; set; }
     public DateTime? ModifiedDate { get; set; }
